+++ src/index.ts
@@ -3,5 +3,12 @@
 export * from './server';
 export * from './tools/common';
 export * from './tools/logger';
-export { enrollUser, getAccountId, getConfiguration, getWallet, isWalletExists } from './tools/util';
-
+export {
+  createNewUser,
+  getAccountId,
+  getConfiguration,
+  getConnectedGateway,
+  getWallet,
+  isWalletExists,
+  processTransaction,
+} from './tools/util';