+++ src/routes.ts
@@ -1,8 +1,8 @@
 import bodyParser from 'body-parser';
 import * as Express from 'express';
-import _ from 'lodash';
 import * as swaggerUI from 'swagger-ui-express';
 import { errorHandler, handleRequest, validateInputParameters, validateResponse } from './handler';
+import { initialize } from './init';
 import { APIDefinition } from './tools/apidefinition';
 import { ChaincodeInventory } from './tools/chaincodeinventory';
 
@@ -13,13 +13,13 @@ import { ChaincodeInventory } from './tools/chaincodeinventory';
  *
  */
 
-export default async function(gatewayExpressApp: Express.Application) {
+export default async function (gatewayExpressApp: Express.Application) {
   gatewayExpressApp.use(bodyParser.json());
   gatewayExpressApp.use(bodyParser.urlencoded({ extended: true }));
   await APIDefinition.createInstance();
   ChaincodeInventory.createInstance();
   // Initialize needed accounts
-  // initialize()
+  initialize();
 
   // OpenAPI documentaion
   gatewayExpressApp.use('/api-docs', swaggerUI.serve);