+++ src/server.ts
@@ -9,10 +9,8 @@ import { Config } from './tools/config';
 import { getLogger } from './tools/logger';
 import * as util from './tools/util';
 
-
 // tslint:disable-next-line: no-var-requires
 const gateway = require('express-gateway');
-
 
 const logger = getLogger(__filename);
 
@@ -21,9 +19,7 @@ const logger = getLogger(__filename);
  *
  */
 function startGateway() {
-  gateway()
-    .load(Config.getConfigItem("config_dir"))
-    .run();
+  gateway().load(Config.getConfigItem('config_dir')).run();
 }
 
 /**
@@ -31,8 +27,10 @@ function startGateway() {
  *
  */
 function startGatewayBackend() {
-  const gw_config = yaml.safeLoad(fs.readFileSync(path.join(Config.getConfigItem("config_dir"), 'gateway.config.yml'), 'utf8'));
-  const _url = new url.URL(gw_config.serviceEndpoints.fabric.url);
+  const gw_config = yaml.load(
+    fs.readFileSync(path.join(Config.getConfigItem('config_dir'), 'gateway.config.yml'), 'utf8')
+  );
+  const _url = new url.URL(gw_config['serviceEndpoints']['fabric']['url']);
   const be = express();
   routes(be);
   be.listen(_url.port, () => logger.info(`Angus router listening on port ${_url.port}`));
@@ -41,7 +39,7 @@ function startGatewayBackend() {
 export class AngusRouter {
   constructor(configFile?: string) {
     // Create configuration settings
-    Config.createInstance(process.env.CONFIG_FILE || configFile );
+    Config.createInstance(process.env.CONFIG_FILE || configFile);
   }
 
   start() {
@@ -50,8 +48,4 @@ export class AngusRouter {
     // Start Angus Router
     setImmediate(() => startGatewayBackend());
   }
-}
-
-export async function processTransaction(params: FabricService, config?: FabricConfig): Promise<any> {
-  return await util.processTransaction(params, config);
 }