+++ src/tools/chaincodeinventory.ts
@@ -12,20 +12,18 @@ export namespace ChaincodeInventory {
   let _inventory: FabricConfig[];
 
   export function createInstance() {
-
     try {
       logger.debug('Creating chaincode inventory...');
-      const _inv = yaml.safeLoad(fs.readFileSync(Config.getConfigItem("inventory_file"), 'utf8'));
-      _inventory = _inv.ccMethods;
-  
-      Object.values(_inv.ccMethods).forEach((_item: any) => {
+      const _inv = yaml.load(fs.readFileSync(Config.getConfigItem('inventory_file'), 'utf8'));
+      _inventory = _inv['ccMethods'];
+
+      Object.values(_inventory).forEach((_item: any) => {
         const _invItem: FabricConfig = _item;
         // This transformation is needed because of YAML label substitution
         _invItem.channelName = _item.channelName.name;
         _inventory.push(_invItem);
       });
       logger.debug('Chaincode inventory created.');
-  
     } catch (error) {
       logger.error(error);
     }